Removed occurences of HoneyDew in application and repaced with getText(R.string.app_name)

+++ app/src/main/java/com/ajbtechnologies/dataaccess/DbHelperImpl.java
@@ -979,7 +979,7 @@ public void deleteList(BasicList list)  {
 			for(Map.Entry entry : listPositions.entrySet()) {
 				int update = 0;
 				UpdateBuilder<BasicList, Integer> updateBuilder = getListDao().updateBuilder();
-				updateBuilder.updateColumnValue(BasicList.ROW_NUM_CMN, (Integer)entry.getValue());
+				updateBuilder.updateColumnValue(BasicList.ROW_NUM_CMN, entry.getValue());
 				updateBuilder.where().idEq((Integer)entry.getKey());
 				update = updateBuilder.update();
 			}
